+++ plugins/code-enhancers-and-utils/all-projects/main/kotlin/com/javiersc/gradle/plugins/all/projects/install/pre/commit/InstallPreCommitTask.kt
@@ -49,8 +49,4 @@ internal fun InstallPreCommitTask.createInstallPreCommitGradleTask() {
 }
 
 internal val Project.preCommitFile: File
-    get() =
-        file("${rootProject.rootDir}/.git/hooks/pre-commit").apply {
-            parentFile.mkdirs()
-            createNewFile()
-        }
+    get() = file("${rootProject.rootDir}/.git/hooks/pre-commit").apply { parentFile.mkdirs() }