+++ plugins/code-enhancers-and-utils/all-projects/main/kotlin/com/javiersc/gradle/plugins/all/projects/install/pre/commit/WriteFilePreCommitTask.kt
@@ -18,7 +18,7 @@ open class WriteFilePreCommitTask : DefaultTask() {
 
     @TaskAction
     fun writeFile() {
-        val currentPreCommitText = preCommitFile.readText()
+        val currentPreCommitText = preCommitFile.apply { createNewFile() }.readText()
 
         val preCommitText: String =
             installPreCommitsDir