+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ### Updated
 
-- `com.diffplug.spotless:spotless-plugin-gradle -> 6.1.1`
+- `com.diffplug.spotless:spotless-plugin-gradle -> 6.1.2`
 - `com.github.ben-manes:gradle-versions-plugin -> 0.41.0`
 - `com.javiersc.gradle-plugins:all-plugins -> 0.1.0-rc.39`